chore: finishing touches

+++ CollectR.Application/Features/Categories/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using CollectR.Application.Abstractions;
 using CollectR.Application.Common;
 using CollectR.Application.Contracts.Persistence;
+using CollectR.Domain;
 
 namespace CollectR.Application.Features.Categories.Commands.DeleteCategory;
 
@@ -12,11 +13,16 @@ internal sealed class DeleteCategoryCommandHandler(ICategoryRepository categoryR
         CancellationToken cancellationToken
     )
     {
-        var category = await categoryRepository.GetByIdAsync(request.Id);
+        var category = await categoryRepository.GetWithDetailsAsync(request.Id);
 
         if (category is null)
         {
             return EntityErrors.NotFound(request.Id);
+        }
+
+        if (category.Collectibles.Count > 0)
+        {
+            return EntityErrors.HasAssignedEntities(nameof(Category), request.Id);
         }
 
         await categoryRepository.DeleteAsync(request.Id);
